頁面架構修改 v0.3 b1002 Felix
Ver0.3 b1002 Felix upload
1.只允許直向layout
2.Activity 結束使用 finish()
3.歡迎頁面削除Button,改為五秒後更換為主頁面layout
4.歡迎頁面削除java class,只留純layout ,更名為Welcome
5.發起頁面改為Start
6.設定頁面改為Setup
7.訊息頁面改為Message
