+++ app/src/main/res/layout/activity_main.xml
@@ -4,94 +4,123 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.student.cup.MainActivity"
+    tools:context="com.example.student.cup.Main2Activity"
+    android:background="@drawable/latte"
     android:orientation="vertical"
-    android:background="@drawable/milktea">
+    >
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="90dp"
+        android:orientation="horizontal"
+
+        >
 
 
 
+        <ImageView
 
-        <RelativeLayout
-            android:layout_width="match_parent"
-            android:layout_height="25dp"
-            android:padding="3dp"
-            >
-
-                <TextView
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:textSize="15dp"
-                    android:textColor="#FFFFFF"
-                    android:text="made in Taiwan"
-                    android:layout_alignParentLeft="true"
-                    android:layout_alignParentBottom="true"
-
-                    />
-
-
-
-
-        </RelativeLayout>
-
-
-
-
-        <TextView
-
-
-            android:id="@+id/textView"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-
-            android:layout_marginTop="190dp"
-            android:text="台灣Cup"
-            android:textStyle="italic"
-            android:layout_gravity="center"
-            android:textColor="#FFFFFF"
-            android:textSize="70dp" />
-
-
-        <Button
-
-            android:id="@+id/next"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:onClick="next"
-            android:layout_gravity="center"
-            android:layout_marginTop="50dp"
-            android:background="@drawable/button"
-            android:text="同意條款"
-            android:textColor="#FFFFFF"
-            android:textSize="18dp"
-
+            android:id="@+id/user"
+            android:layout_width="90dp"
+            android:layout_height="90dp"
+            android:src="@drawable/user"
 
             />
 
-        <RelativeLayout
+    <LinearLayout
+        android:layout_width="230dp"
+        android:layout_height="match_parent"
+        android:orientation="vertical"
+
+        >
+
+
+        <TextView
+            android:id="@+id/name"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="黃浩庭"
+            android:textSize="23dp"
+            android:textColor="#FFFFFF"
+            android:layout_marginLeft="7dp"
+            android:layout_marginTop="1dp"
+
+
+            />
+        <EditText
             android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:padding="3dp"
-            >
-                <TextView
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:textSize="15dp"
-                    android:textColor="#FFFFFF"
-                    android:text="since.2017.9.29"
-                    android:layout_alignParentRight="true"
-                    android:layout_alignParentBottom="true"
+            android:layout_height="wrap_content"
+            android:text="你在想什麼?這裡我設定最多打24個字剛好兩行國字"
+            android:textColor="#FFFFFF"
+            android:textSize="18dp"
+            android:maxLength="24"
 
-                    />
+            />
 
 
 
 
+    </LinearLayout>
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:orientation="vertical"
+        >
+
+        <ImageButton
+
+            android:id="@+id/settinglink"
+            android:layout_width="40dp"
+            android:layout_height="40dp"
+            android:layout_gravity="center"
+            android:background="@drawable/setting2"
+            android:layout_marginTop="5dp"
+            android:onClick="settinglink"
+
+            />
+
+        <ImageButton
+            android:id="@+id/messagelink"
+            android:layout_width="40dp"
+            android:layout_height="40dp"
+            android:layout_gravity="center"
+            android:background="@drawable/messagew"
+            android:layout_marginTop="5dp"
+            android:onClick="messagelink"
+            />
+
+    </LinearLayout>
+
+    </LinearLayout>
+
+    <TextView
+        android:id="@+id/online"
+        android:layout_width="match_parent"
+        android:layout_height="330dp"
+        android:layout_marginTop="8dp"
+        android:background="@android:drawable/alert_light_frame"
+        android:text="這裡寫訂購中的列表\n例如說 : \n某某人:目前訂購奶茶中\n然後最右邊有按鈕加入訂購，之類的"
+        android:textSize="30dp"
 
 
-        </RelativeLayout>
+
+        />
+
+    <Button
 
 
+        android:id="@+id/start"
+        android:layout_width="200dp"
+        android:layout_height="80dp"
+        android:layout_gravity="center"
+        android:layout_marginTop="40dp"
+        android:background="@drawable/button"
+        android:text="開始訂購"
+        android:textColor="#FFFFFF"
+        android:onClick="start"
 
+        android:textSize="30dp" />
 
 
 </LinearLayout>
