--- app/src/main/res/layout/activity_main2.xml
@@ -1,126 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.example.student.cup.Main2Activity"
-    android:background="@drawable/latte"
-    android:orientation="vertical"
-    >
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="90dp"
-        android:orientation="horizontal"
-
-        >
-
-
-
-        <ImageView
-
-            android:id="@+id/user"
-            android:layout_width="90dp"
-            android:layout_height="90dp"
-            android:src="@drawable/user"
-
-            />
-
-    <LinearLayout
-        android:layout_width="230dp"
-        android:layout_height="match_parent"
-        android:orientation="vertical"
-
-        >
-
-
-        <TextView
-            android:id="@+id/name"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="黃浩庭"
-            android:textSize="23dp"
-            android:textColor="#FFFFFF"
-            android:layout_marginLeft="7dp"
-            android:layout_marginTop="1dp"
-
-
-            />
-        <EditText
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="你在想什麼?這裡我設定最多打24個字剛好兩行國字"
-            android:textColor="#FFFFFF"
-            android:textSize="18dp"
-            android:maxLength="24"
-
-            />
-
-
-
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical"
-        >
-
-        <ImageButton
-
-            android:id="@+id/settinglink"
-            android:layout_width="40dp"
-            android:layout_height="40dp"
-            android:layout_gravity="center"
-            android:background="@drawable/setting2"
-            android:layout_marginTop="5dp"
-            android:onClick="settinglink"
-
-            />
-
-        <ImageButton
-            android:id="@+id/messagelink"
-            android:layout_width="40dp"
-            android:layout_height="40dp"
-            android:layout_gravity="center"
-            android:background="@drawable/messagew"
-            android:layout_marginTop="5dp"
-            android:onClick="messagelink"
-            />
-
-    </LinearLayout>
-
-    </LinearLayout>
-
-    <TextView
-        android:id="@+id/online"
-        android:layout_width="match_parent"
-        android:layout_height="330dp"
-        android:layout_marginTop="8dp"
-        android:background="@android:drawable/alert_light_frame"
-        android:text="這裡寫訂購中的列表\n例如說 : \n某某人:目前訂購奶茶中\n然後最右邊有按鈕加入訂購，之類的"
-        android:textSize="30dp"
-
-
-
-        />
-
-    <Button
-
-
-        android:id="@+id/start"
-        android:layout_width="200dp"
-        android:layout_height="80dp"
-        android:layout_gravity="center"
-        android:layout_marginTop="40dp"
-        android:background="@drawable/button"
-        android:text="開始訂購"
-        android:textColor="#FFFFFF"
-        android:onClick="start"
-
-        android:textSize="30dp" />
-
-
-</LinearLayout>
