+++ app/src/main/res/layout/activity_message.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.student.cup.message"
+    tools:context="com.example.student.cup.MessageActivity"
     android:orientation="vertical"
     android:background="@drawable/latte"
