+++ app/src/main/res/layout/activity_open.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.student.cup.Main3Activity"
+    tools:context="com.example.student.cup.OpenActivity"
     android:orientation="vertical"
     android:background="@drawable/latte"
