+++ app/src/main/res/layout/layout_welcome.xml
@@ -1,0 +1,97 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context="com.example.student.cup.MainActivity"
+    android:orientation="vertical"
+    android:background="@drawable/milktea">
+
+
+
+
+        <RelativeLayout
+            android:layout_width="match_parent"
+            android:layout_height="25dp"
+            android:padding="3dp"
+            >
+
+                <TextView
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:textSize="15dp"
+                    android:textColor="#FFFFFF"
+                    android:text="made in Taiwan"
+                    android:layout_alignParentLeft="true"
+                    android:layout_alignParentBottom="true"
+
+                    />
+
+
+
+
+        </RelativeLayout>
+
+
+
+
+        <TextView
+
+
+            android:id="@+id/textView"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+
+            android:layout_marginTop="190dp"
+            android:text="台灣Cup"
+            android:textStyle="italic"
+            android:layout_gravity="center"
+            android:textColor="#FFFFFF"
+            android:textSize="70dp" />
+
+<!--
+        <Button
+
+            android:id="@+id/next"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center"
+            android:layout_marginTop="50dp"
+            android:background="@drawable/button"
+            android:onClick="next"
+            android:text="同意條款"
+            android:textColor="#FFFFFF"
+            android:textSize="18dp"
+
+
+            />
+-->
+        <RelativeLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:padding="3dp"
+            >
+                <TextView
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:textSize="15dp"
+                    android:textColor="#FFFFFF"
+                    android:text="since.2017.9.29"
+                    android:layout_alignParentRight="true"
+                    android:layout_alignParentBottom="true"
+
+                    />
+
+
+
+
+
+
+        </RelativeLayout>
+
+
+
+
+
+</LinearLayout>
